+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # ActFramework Change Log
 
 **1.8.31**
+* Enable turn off QRCode/BarCode rendering based on Accept header #1269
 * ParamValueBinder - allow bind to ActionContext renderArgs #1267
 * CLI - add shortcut into the command help list #1266
 * No response on certain error scenario #1264